requirements.txt
# PhotoTrick Python Dependencies
# 用于OCR服务和文件转换

# PaddleOCR
paddlepaddle>=2.5.0
paddleocr>=2.7.0

# Web框架
flask>=2.0.0

# 图片处理
pillow>=9.0.0
pillow-heif>=0.10.0  # HEIC格式支持

# PDF处理
pdf2image>=1.16.0
pypdf>=3.0.0

# Word/Excel处理
python-docx>=0.8.11
openpyxl>=3.1.0

# HTML解析（表格识别）
beautifulsoup4>=4.12.0
lxml>=4.9.0

# 可选：OFD格式支持
# ofdrw>=0.1.0  # 需要从特定源安装
